chore: Add externalfuture for FailingCallable
